- [리뷰 반영] UIView extension -> Undoable 프로토콜 방식으로 변경
